MAYA-121735: Fix zero tangents
While investigating the normal export for a cube I noticed that 2 of the
faces were rendering black due to tangents being zero. Augmented the
already arbitrary fix to at least not have buggy faces. To be improved
further with a proper geometry shader that uses the UV stream when
available.

Also added MaterialX shims for Lambert, Blinn, and Phong allowing
import/export and minimal VP2 rendering.